# docs/swagger.v1.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Mini E-Commerce API",
    "version": "v1"
  },
  "paths": {
...
    "/api/autenticacion/registro": {
      "post": {
        "tags": [
          "Autenticacion"
        ],
        "summary": "Registra un usuario nuevo.",
        "description": "Crea un usuario con rol <b>User</b>, genera <b>AccessToken (JWT)</b> y <b>RefreshToken</b>.\r\n\r\n\r\n<b>Seguridad:</b>\r\n- No loguear el token.\r\n- El refresh token se guarda en base de datos.",
        "requestBody": {
          "description": "Datos del registro (email, nombre y contraseña).",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistroRequestDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistroRequestDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RegistroRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuario registrado correctamente. Devuelve AccessToken + RefreshToken.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AutenticacionResponseDtoApiResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AutenticacionResponseDtoApiResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AutenticacionResponseDtoApiResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Error de validación o email ya registrado.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/autenticacion/login": {
      "post": {
        "tags": [
          "Autenticacion"
        ],
        "summary": "Inicia sesión.",
        "description": "Valida credenciales y devuelve:\r\n- <b>AccessToken</b> (JWT)\r\n- <b>RefreshToken</b>\r\n\r\n\r\nSi las credenciales son incorrectas, devuelve un error estándar.",
        "requestBody": {
          "description": "Credenciales del usuario.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequestDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequestDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login exitoso. Devuelve AccessToken + RefreshToken.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AutenticacionResponseDtoApiResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AutenticacionResponseDtoApiResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AutenticacionResponseDtoApiResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Credenciales inválidas.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/autenticacion/refresh-token": {
      "post": {
        "tags": [
          "Autenticacion"
        ],
        "summary": "Renueva el access token usando refresh token.",
        "description": "Rotación de refresh token:\r\n- El token anterior pasa a estado <b>Used/Revoked</b>.\r\n- Se emite un nuevo RefreshToken.\r\n\r\n\r\n\r\nEstados de error comunes:\r\n- Token no existe\r\n- Token expirado\r\n- Token revocado/usado\r\n- Usuario no existe asociado al token",
        "requestBody": {
          "description": "Refresh token válido.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenRenovarDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenRenovarDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenRenovarDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token renovado. Devuelve AccessToken + RefreshToken nuevo.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AutenticacionResponseDtoApiResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AutenticacionResponseDtoApiResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AutenticacionResponseDtoApiResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Token inválido (no existe, expirado, revocado, usado).",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/autenticacion/logout": {
      "post": {
        "tags": [
          "Autenticacion"
        ],
        "summary": "Cierra sesión (logout).",
        "description": "Revoca el refresh token enviado:\r\n- Marca <b>IsUsed=true</b>\r\n- Guarda <b>RevokedAt</b>\r\n\r\n\r\nNota: No invalida el AccessToken inmediatamente (por diseño de JWT).\r\nLa sesión queda cerrada porque el refresh token deja de ser válido.",
        "requestBody": {
          "description": "Refresh token a revocar.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenRenovarDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenRenovarDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenRenovarDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Logout exitoso. Token revocado.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Token inválido (no existe, expirado, revocado, usado).",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiErrorDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "nullable": true
          },
          "message": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "AutenticacionResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "nullable": true
          },
          "refreshToken": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "AutenticacionResponseDtoApiResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "value": {
            "$ref": "#/components/schemas/AutenticacionResponseDto"
          },
          "error": {
            "$ref": "#/components/schemas/ApiErrorDto"
          }
        },
        "additionalProperties": false
      },
...
      "LoginRequestDto": {
        "required": [
          "email",
          "password"
        ],
        "type": "object",
        "properties": {
          "email": {
            "minLength": 1,
            "type": "string",
            "format": "email"
          },
          "password": {
            "minLength": 1,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "ObjectApiResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "value": {
            "nullable": true
          },
          "error": {
            "$ref": "#/components/schemas/ApiErrorDto"
          }
        },
        "additionalProperties": false
      },
...
      "RefreshTokenRenovarDto": {
        "required": [
          "token"
        ],
        "type": "object",
        "properties": {
          "token": {
            "minLength": 1,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "RegistroRequestDto": {
        "required": [
          "email",
          "nombre",
          "password"
        ],
        "type": "object",
        "properties": {
          "email": {
            "minLength": 1,
            "type": "string",
            "format": "email"
          },
          "nombre": {
            "minLength": 2,
            "type": "string"
          },
          "password": {
            "minLength": 8,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Ingrese: Bearer {tu_token}",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}